Moved to snacks.picker from telescope

+++ dot_config/zsh/functions/plugins_management.zsh
@@ -1,6 +1,5 @@
 #!/usr/bin/env zsh
 
-
 function plugin-load {
     local plugin repo commitsha plugdir initfile initfiles=() clone_args=()
     for plugin in $@; do
@@ -25,14 +24,13 @@ function plugin-load {
         fi
         if [[ ! -e $initfile ]]; then
             initfiles=($plugdir/*.{plugin.zsh,zsh-theme,zsh,sh}(N))
-            (( $#initfiles )) || { echo >&2 "No init file found '$repo'." && continue }
+            (($#initfiles)) || { echo >&2 "No init file found '$repo'." && continue; }
             ln -sf $initfiles[1] $initfile
         fi
         fpath+=$plugdir
-        (( $+functions[zsh-defer] )) && zsh-defer . $initfile || . $initfile
+        (($+functions[zsh - defer])) && zsh-defer . $initfile || . $initfile
     done
 }
-
 
 function plugin-compile {
     autoload -U zrecompile
@@ -42,24 +40,24 @@ function plugin-compile {
     done
 }
 
-
 function plugin-update {
     local update_interval=$1
-    local min_time_diff=$(( 24 * 60 * 60 * update_interval ))
+    local min_time_diff=$((24 * 60 * 60 * update_interval))
     local current_time=$(date +%s)
     local time_file=$_zplugindir/last_update_secs.tmp
 
-    local last_time=$(cat $time_file 2> /dev/null || echo 0)
-    local time_diff=$(( current_time - last_time ))
+    local last_time=$(cat $time_file 2>/dev/null || echo 0)
+    local time_diff=$((current_time - last_time))
 
-    if (( time_diff >= min_time_diff )); then
-        rm -rf $_zplugindir 2>/dev/null
-        mkdir -p $_zplugindir
-        echo $current_time > $time_file
-        echo 'Updating ZSH plugins...'
+    if ((time_diff >= min_time_diff)); then
+        if ping -c 1 -W 2 google.com &>/dev/null; then
+            rm -rf $_zplugindir 2>/dev/null
+            mkdir -p $_zplugindir
+            echo $current_time >$time_file
+            echo 'Updating ZSH plugins...'
+        fi
     fi
 }
-
 
 function fast-theme-generate() {
     mkdir -p $FAST_WORK_DIR
